Update fx_feat.py
Fixed an issue where the FISM2 netcdf file didn't have good data after 2022. Dropped data after this.

Added append_feat which will append data to the feature array up to an end date if the data wasn't there.

Added append_feat to load_feat.
If the difference between the end of the loaded feature data and the end date requested is larger then a day, then append is called in order to get the rest of the required feature data and append it to the feature dataframe.